Fix bug in INC instructions where zero flag was not set
The operation returns its result as an integer larger than 8 bit, which
means it can't overflow from 0xFF to 0x00, so the zero flag would not be
set. By casting to uint8_t the overflow is recognized, since the value
wraps around to 0.


Former-commit-id: ba86416312f96b04e81a155e003d2c092d276859

+++ src/game-boy-emulator/cpu.cpp
@@ -583,7 +583,8 @@ void Cpu::instructionINCDEC(opcodes::Instruction instruction) {
     default:
         auto was_hc = internal::was_half_carry(value_original, 1, operation);
         set_half_carry_flag(was_hc);
-        set_zero_flag(operation(value_original, 1) == 0);
+        // Truncate to 8 bits to catch overflows/underflows
+        set_zero_flag(static_cast<uint8_t>(operation(value_original, 1)) == 0);
         set_subtract_flag(instruction.instruction_type == opcodes::InstructionType::INC
                               ? BitValues::Inactive
                               : BitValues::Active);
